implemented toggleable VRAM viewer

+++ src/core/sink.rs
@@ -1,5 +1,5 @@
 pub struct VideoSink {
-	inner: Option<Vec<u8>>
+	inner: Option<Vec<u32>>
 }
 
 impl VideoSink {
@@ -8,10 +8,10 @@ impl VideoSink {
 			inner: None
 		}
 	}
-	pub fn consume(self) -> Option<Vec<u8>> {
+	pub fn consume(self) -> Option<Vec<u32>> {
 		self.inner
 	}
-	pub fn append(&mut self, value: Vec<u8>) {
+	pub fn append(&mut self, value: Vec<u32>) {
 		self.inner = Some(value);
 	}
 }
